+++ src/apps/ScopeChangeRequest/Api/PCS/getSystems.ts
@@ -1,5 +1,5 @@
 import { httpClient } from '../../../../Core/Client/Functions/HttpClient';
-import { System } from '../Search/PCS/Types/system';
+import { System } from '../../Types/ProCoSys/system';
 
 export async function getSystems(): Promise<System[]> {
     const { procosys } = httpClient();
