--- src/apps/ScopeChangeRequest/Api/Search/PCS/Types/area.ts
@@ -1,5 +1,0 @@
-export interface Area {
-    Code: string;
-    Description: string;
-    Id: number;
-}
