--- src/apps/ScopeChangeRequest/Api/Search/PCS/Types/commpkg.ts
@@ -1,5 +1,0 @@
-export interface CommPkg {
-    Id: string;
-    CommPkgNo: string;
-    Description: string;
-}
