--- src/apps/ScopeChangeRequest/Api/Search/PCS/Types/functionalRole.ts
@@ -1,7 +1,0 @@
-export interface FunctionalRole {
-    Code: string;
-    LibraryType__Id: string;
-    Path: string;
-    Id: number;
-    Description: string;
-}
