--- src/apps/ScopeChangeRequest/Api/Search/PCS/Types/index.ts
@@ -1,7 +1,0 @@
-export * from './tag';
-export * from './commpkg';
-export * from './functionalRole';
-export * from './person';
-export * from './query';
-export * from './searchStructure';
-export * from './system';
