--- src/apps/ScopeChangeRequest/Api/Search/PCS/Types/tag.ts
@@ -1,5 +1,0 @@
-export interface Tag {
-    TagNo: string;
-    Id: string;
-    Description: string;
-}
