--- src/apps/ScopeChangeRequest/Api/Search/PCS/index.ts
@@ -1,1 +1,0 @@
-export * from './Types';
