+++ src/apps/ScopeChangeRequest/Api/Search/PCS/searchDCN.ts
@@ -1,8 +1,8 @@
 import { isProduction } from '../../../../../Core/Client/Functions';
 import { HttpClient } from '../../../../../Core/httpClient/src';
 import { TypedSelectOption } from '../searchType';
-import { DCN } from './Types/dcn';
-import { PCSStructure } from './Types/searchStructure';
+import { DCN } from '../../../Types/ProCoSys/dcn';
+import { PCSStructure } from './searchStructure';
 
 export const searchDCN = async (
     searchString: string,
